requirements.txt
arrow==0.15.2
beautifulsoup4==4.8.0
Faker==2.0.2
matplotlib==3.2.1
mysqlclient==1.4.4
numpy==1.17.2
pandas==0.25.1
PyYAML==5.3
requests==2.22.0
selenium==3.141.0
SQLAlchemy==1.3.8
statsmodels==0.11.1
tenacity==6.0.0
wrapt==1.11.2
